+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ⭐ 0.0.2 | 2024-03-20
 - Add `mrms.timeseries` tools. For larger datasets, these run faster than using `xr.open_mfdataset`.
 - Add docstrings to functions.
+- Build timeseries using polygons.
 
 ## Previous notes: